Separate probabilistic analysis and metrics by files.
